Establish basic style analyzer rules (#3759)
* Fix analyzer nullable error

can't disable them in CA1305 xdd

* .editorconfig -> .globalconfig

* Add basic set of style analyzer rules

* dotnet format output

* Some additional fixes

* Apply manual simplification

* Transform some enum comparisons

* fix typo

* add TODO

+++ src/BizHawk.Emulation.Cores/CPUs/68000/Instructions/DataMovement.cs
@@ -508,7 +508,7 @@ namespace BizHawk.Emulation.Cores.Components.M68000
 			string address = DisassembleAddress(mode, reg, ref pc);
 
 			info.Mnemonic = size == 0 ? "movem.w" : "movem.l";
-			info.Args = DisassembleRegisterList0((uint)registers) + ", " + address;
+			info.Args = DisassembleRegisterList0(registers) + ", " + address;
 			info.Length = pc - info.PC;
 		}
