+++ src/BizHawk.Emulation.Cores/CPUs/68000/Instructions/Supervisor.cs
@@ -29,7 +29,7 @@ namespace BizHawk.Emulation.Cores.Components.M68000
 		{
 			int mode = (op >> 3) & 7;
 			int reg = (op >> 0) & 7;
-			WriteValueW(mode, reg, (short)SR);
+			WriteValueW(mode, reg, SR);
 			PendingCycles -= (mode == 0) ? 6 : 8 + EACyclesBW[mode, reg];
 		}
 
@@ -151,7 +151,7 @@ namespace BizHawk.Emulation.Cores.Components.M68000
 
 		private void TrapVector(int vector)
 		{
-			short sr = (short)SR;        // capture current SR.
+			short sr = SR;        // capture current SR.
 			S = true;                    // switch to supervisor mode, if not already in it.
 			A[7].s32 -= 4;               // Push PC on stack
 			WriteLong(A[7].s32, PC);
